requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
multidict==6.0.5
openai==1.14.0
pinecone-client==3.1.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
rsa==4.9
simple-websocket==1.0.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.4
